Add methods for Curve25519 point encoding.
This safely wraps the unsafe C calls and gives us a destination for
their upcoming oxidization. We don't include a slice-mutating point
encoding method because its measured benefit was negligible.

Note that we stop using the `Scalar` alias: this is intentional.
Although `SCALAR_LEN` == `ELEM_LEN`, the byte buffers in question are
not encoded scalars, but the 31-bit encodings of the Y coordinate of a
curve point, enriched with a sign bit.

I agree to license my contributions to each file under the terms given
at the top of each file I changed.